requirements.txt
asgiref==3.4.1
coverage==6.1.1
Django==3.2.8
django-dotenv==1.4.2
djangorestframework==3.12.4
Pillow==8.4.0
psycopg2==2.9.1
PyJWT==2.3.0
pytz==2021.3
sqlparse==0.4.2
